ADD front end of the app.
Create functions to parse the request, process it and the app works correctly.
Is necesary to update the info provided by the front end and implement tests and docs.
